[Feature] Better translate test (#39) (#47)
Translate test: small improvements

- Parametrize perturbation upon failure
- Refactor error folder to be `pwd` based
- Fix GPU translate unable to dump error `nc`
- Fix mixmatch precision on translate test
- Update README.md

Test fix:
- Orchestrate YPPM for translate purposes

Misc:
- Fix bad logger formatting on DaCeProgress

+++ fv3core/tests/savepoint/translate/translate_yppm.py
@@ -1,6 +1,7 @@
 import pace.dsl
 import pace.dsl.gt4py_utils as utils
 import pace.util
+from pace.dsl.typing import Float
 from pace.fv3core.stencils import yppm
 from pace.fv3core.testing import TranslateDycoreFortranData2Py
 from pace.stencils.testing import TranslateGrid
@@ -40,7 +41,9 @@ class TranslateYPPM(TranslateDycoreFortranData2Py):
         self.ivars(inputs)
         self.make_storage_data_input_vars(inputs)
         inputs["flux"] = utils.make_storage_from_shape(
-            inputs["q"].shape, backend=self.stencil_factory.backend
+            inputs["q"].shape,
+            backend=self.stencil_factory.backend,
+            dtype=Float,
         )
 
     def compute(self, inputs):
